% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaprime.R
\name{betaprime}
\alias{betaprime}
\alias{dbetaprime}
\alias{pbetaprime}
\alias{qbetaprime}
\alias{rbetaprime}
\title{Beta prime distribution}
\usage{
dbetaprime(x, shape1, shape2, log = FALSE)

pbetaprime(q, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

qbetaprime(p, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

rbetaprime(n, shape1, shape2)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{shape1, shape2}{non-negative shape parameters of the corresponding Beta distribution}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/ densities \eqn{p} are returned as \eqn{\log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \le x]}, otherwise \eqn{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of random values to return.}
}
\value{
\code{dbetaprime} gives the density, \code{pbetaprime} gives the distribution function, \code{qbetaprime} gives the quantile function, and \code{rbetaprime} generates random deviates.
}
\description{
Density, distribution function, quantile function, and random generation for
the Beta prime distribution.
}
\details{
This implementation allows for automatic differentiation with \code{RTMB}.

If \eqn{X \sim \text{Beta}(\alpha, \beta)}, then \eqn{\frac{X}{1-X} \sim \text{Betaprime}(\alpha, \beta)}
}
\examples{
x <- rbetaprime(1, 2, 1)
d <- dbetaprime(x, 2, 1)
p <- pbetaprime(x, 2, 1)
q <- qbetaprime(p, 2, 1)
}
