% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{erf}
\alias{erf}
\alias{erfc}
\title{AD-compatible error function and complementary error function}
\usage{
erf(x)

erfc(x)
}
\arguments{
\item{x}{vector of evaluation points}
}
\value{
\code{erf(x)} returns the error function and \code{erfc(x)} returns the complementary error function.
}
\description{
AD-compatible error function and complementary error function
}
\examples{
erf(1)
erfc(1)
}
