% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invchisq.R
\name{invchisq}
\alias{invchisq}
\alias{dinvchisq}
\alias{pinvchisq}
\alias{qinvchisq}
\alias{rinvchisq}
\title{Inverse Chi-squared distribution}
\usage{
dinvchisq(x, df, scale = 1/df, log = FALSE)

pinvchisq(q, df, scale = 1/df, lower.tail = TRUE, log.p = FALSE)

qinvchisq(p, df, scale = 1/df, lower.tail = TRUE, log.p = FALSE)

rinvchisq(n, df, scale = 1/df)
}
\arguments{
\item{x, q}{vector of quantiles, must be positive.}

\item{df}{degrees of freedom (\eqn{\nu > 0})}

\item{scale}{optional positive scale parameter. Default value of \code{1/df} corresponds to standard inverse gamma}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/densities are returned as \eqn{\log(p)}.}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of random values to return}
}
\value{
\code{dinvchisq} gives the density, \code{pinvchisq} gives the distribution function,
\code{qinvchisq} gives the quantile function, and \code{rinvchisq} generates random deviates.
}
\description{
Density, distribution function, quantile function, and random generation
for the inverse Chi-squared distribution.
}
\details{
If \eqn{X \sim \text{Chisq}(\nu)}, then \eqn{1/X \sim \text{invChisq}(\nu)}.

The inverse Chi-squared distribution with \eqn{\nu} degrees of freedom has
density
\deqn{f(x) = \frac{(\nu/2)^{\nu/2}}{\Gamma(\nu/2)} x^{-(\nu/2+1)} \exp(-\nu/(2x)), \quad x>0.}

This implementation of \code{dinvchisq}, \code{pinvchisq}, and \code{qinvchisq} allows for automatic differentiation with \code{RTMB}.
}
\examples{
x <- rinvchisq(1, df = 5)
d <- dinvchisq(x, df = 5)
p <- pinvchisq(x, df = 5)
q <- qinvchisq(p, df = 5)
}
