% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumar.R
\name{kumar}
\alias{kumar}
\alias{dkumar}
\alias{pkumar}
\alias{qkumar}
\alias{rkumar}
\title{Kumaraswamy distribution}
\usage{
dkumar(x, a, b, log = FALSE)

pkumar(q, a, b, lower.tail = TRUE, log.p = FALSE)

qkumar(p, a, b, lower.tail = TRUE, log.p = FALSE)

rkumar(n, a, b)
}
\arguments{
\item{x, q}{vector of quantiles in \eqn{(0,1)}}

\item{a, b}{positive shape parameters}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/ densities \eqn{p} are returned as \eqn{\log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \le x]}, otherwise \eqn{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of random values to return.}
}
\value{
\code{dkumar} gives the density, \code{pkumar} gives the distribution function, \code{qkumar} gives the quantile function, and \code{rkumar} generates random deviates.
}
\description{
Density, distribution function, quantile function, and random generation for
the Kumaraswamy distribution.
}
\examples{
x <- rkumar(1, a = 1, b = 2)
d <- dkumar(x, a = 1, b = 2)
p <- pkumar(x, a = 1, b = 2)
q <- qkumar(p, a = 1, b = 2)
}
