% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcreport.R
\name{mcreport}
\alias{mcreport}
\title{Sample parameters from approximate Gaussian posterior distribution}
\usage{
mcreport(
  obj,
  nSamples = 1000,
  include_random_pars = TRUE,
  report = FALSE,
  Q = NULL,
  ...
)
}
\arguments{
\item{obj}{Optimised \code{RTMB} object generated by \code{\link[RTMB]{MakeADFun}}}

\item{nSamples}{Number of samples to draw}

\item{include_random_pars}{Logical; Should random parameters be included in the output?}

\item{report}{Logical; Should \code{REPORT}ed quantities be sampled as well?
Defaults to \code{FALSE} because this may be slow depending on your model.}

\item{Q}{Optional precalculated sparse precision matrix returned by \code{sdreport(..., getJointPrecision = TRUE)$jointPrecision}.
If not provided, computed internally using \code{sdreport}. Only used for models with random effects.}

\item{...}{For internal use only}
}
\value{
A list structured like the original parameter list used in the \code{\link[RTMB]{MakeADFun}} call (potentially including additional \code{REPORT}ed quantities). Each entry is a list with \code{nSamples} entries.
}
\description{
Efficient Monte Carlo sampling of parameters (and \code{REPORT}ed quantities) from the approximate posterior of an \code{RTMB} model.
See \code{\link[TMB]{sdreport}} for details on posterior variance-covariance in random effects models.
}
\details{
\strong{Caution:}
This has nothing to do with Bayesian posterior sampling.
It is simple Monte Carlo sampling from a Gaussian distribution with mean at the MLE and covariance given by the inverse of the Hessian (for fixed effects models) or the joint precision matrix (for random effects models).
This is a common approach to get an approximate idea of parameter uncertainty around the MLE, but it relies on large-sample asymptotics.

Sampling random effects from their posterior as compared to calculating marginal standard devitions like \code{\link[TMB]{sdreport}} is particularly useful for multidimensional random effects (e.g. for locations \eqn{x} and \eqn{y})
where pointwise confidence intervals (e.g. along a path) based on standard deviations are not possible.
}
\examples{
set.seed(123)

### Example 1: Without random effects ###

# simulate data
x <- rgumbel(100, location = 5, scale = 2)

# negative log-likelihood function
nll <- function(par) {
  getAll(par)
  x <- OBS(x) # mark x as the response
  scale <- exp(log_scale)
  ADREPORT(scale); REPORT(scale)
  -sum(dgumbel(x, loc, scale, log = TRUE))
}

# RTMB AD object
par <- list(loc = 5, log_scale = log(2))
obj <- MakeADFun(nll, par, silent = TRUE)

# model fitting using AD gradient
opt <- nlminb(obj$par, obj$fn, obj$gr)

# sample from asymptotic distribution of the MLE
samples <- mcreport(obj, report = TRUE)
# parameters
mean(unlist(samples$loc)); sd(unlist(samples$loc))
# reported transformed parameters
mean(unlist(samples$scale)); sd(unlist(samples$scale))

# compare to delta method sd from sdreport
sdr <- sdreport(obj)
summary(sdr)



### Example 2: With random effects ###

# simulate random effects
id <- rep(1:10, each = 100) # 10 groups with 100 observations each
true_gamma <- rnorm(10, 0, 2) # random coefficients
true_probs <- plogis(1 + true_gamma) # linear predictor

# simulate Bernoulli data conditional on random coefficients
x <- rbinom(1000, 1, true_probs[id])

# negative log-likelihood function
nll <- function(par) {
  getAll(par)
  x <- OBS(x) # mark x as the response

  # random effect likelihood
  sd_gamma <- exp(log_sd_gamma)
  ADREPORT(sd_gamma); REPORT(sd_gamma)
  nll <- -sum(dnorm(gamma, 0, sd_gamma, log = TRUE))

  # data likelihood
  probs <- plogis(beta0 + gamma) # linear predictor
  ADREPORT(probs); REPORT(probs)
  nll <- nll - sum(dbinom(x, 1, probs[id], log = TRUE))
  return(nll)
}

# RTMB Laplace approximation
par <- list(beta0 = 1, log_sd_gamma = log(2), gamma = rep(0,10))
obj <- MakeADFun(nll, par, random = "gamma", silent = TRUE)

# model fitting using AD gradient
opt <- nlminb(obj$par, obj$fn, obj$gr)

# draw samples
samples <- mcreport(obj, report = TRUE)

# run sdreport
sdr <- sdreport(obj)

# standard deviation using delta method
as.list(sdr, "Std", report = TRUE)$probs

# standard deviation from samples
apply(simplify2array(samples$probs), 1, sd)
}
