% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe.R
\name{pe}
\alias{pe}
\alias{dpe}
\alias{ppe}
\alias{qpe}
\alias{rpe}
\alias{dpe2}
\alias{ppe2}
\alias{qpe2}
\alias{rpe2}
\title{Power Exponential distribution (PE and PE2)}
\usage{
dpe(x, mu = 0, sigma = 1, nu = 2, log = FALSE)

ppe(q, mu = 0, sigma = 1, nu = 2, lower.tail = TRUE, log.p = FALSE)

qpe(p, mu = 0, sigma = 1, nu = 2, lower.tail = TRUE, log.p = FALSE)

rpe(n, mu = 0, sigma = 1, nu = 2)

dpe2(x, mu = 0, sigma = 1, nu = 2, log = FALSE)

ppe2(q, mu = 0, sigma = 1, nu = 2, lower.tail = TRUE, log.p = FALSE)

qpe2(p, mu = 0, sigma = 1, nu = 2, lower.tail = TRUE, log.p = FALSE)

rpe2(n, mu = 0, sigma = 1, nu = 2)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{sigma}{scale parameter, must be positive}

\item{nu}{shape parameter (real)}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/ densities \eqn{p} are returned as \eqn{\log(p)}}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \le x]}, otherwise \eqn{P[X > x]}}

\item{p}{vector of probabilities}

\item{n}{number of random values to return}
}
\value{
\code{dpe} gives the density, \code{ppe} gives the distribution function, \code{qpe} gives the quantile function, and \code{rpe} generates random deviates.
}
\description{
Density, distribution function, quantile function, and random generation for
the Power Exponential distribution (two versions).
}
\details{
This implementation of \code{dpe} and \code{dpe2} allows for automatic differentiation with \code{RTMB} while the other functions are imported from \code{gamlss.dist} package.

For \code{PE}, \code{mu} is the mean and \code{sigma} is the standard deviation while this does not hold for \code{PE2}.

See \code{gamlss.dist::\link[gamlss.dist]{PE}} for more details.
}
\examples{
# PE
x <- rpe(5, mu = 0, sigma = 1, nu = 2)
d <- dpe(x, mu = 0, sigma = 1, nu = 2)
p <- ppe(x, mu = 0, sigma = 1, nu = 2)
q <- qpe(p, mu = 0, sigma = 1, nu = 2)

# PE2
x <- rpe2(5, mu = 0, sigma = 1, nu = 2)
d <- dpe2(x, mu = 0, sigma = 1, nu = 2)
p <- ppe2(x, mu = 0, sigma = 1, nu = 2)
q <- qpe2(p, mu = 0, sigma = 1, nu = 2)
}
