% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmrf.R
\name{rgmrf}
\alias{rgmrf}
\title{Sample from a multivariate Gaussian with a sparse precision matrix}
\usage{
rgmrf(n, mean = 0, Q)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{mean}{Mean vector (or scalar, which will be recycled to match the dimension of \code{Q}).}

\item{Q}{Sparse precision matrix (\eqn{\Sigma^{-1}}).}
}
\value{
A matrix of samples with rows corresponding to samples and columns to dimensions.
}
\description{
Draw samples from a multivariate Gaussian distribution specified by a sparse precision matrix.
This is numerically efficient for high-dimensional but sparse systems.
}
\examples{
rgmrf(3, mean = c(1, 2, 3), Q = Matrix::Diagonal(3))
}
