% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnorm.R
\name{skewnorm}
\alias{skewnorm}
\alias{dskewnorm}
\alias{pskewnorm}
\alias{qskewnorm}
\alias{rskewnorm}
\title{Skew normal distribution}
\usage{
dskewnorm(x, xi = 0, omega = 1, alpha = 0, log = FALSE)

pskewnorm(q, xi = 0, omega = 1, alpha = 0, lower.tail = TRUE, log.p = FALSE)

qskewnorm(p, xi = 0, omega = 1, alpha = 0, lower.tail = TRUE, log.p = FALSE)

rskewnorm(n, xi = 0, omega = 1, alpha = 0)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{xi}{location parameter}

\item{omega}{scale parameter, must be positive.}

\item{alpha}{skewness parameter, +/- \code{Inf} is allowed.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/ densities \eqn{p} are returned as \eqn{\log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \le x]}, otherwise \eqn{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of random values to return}
}
\value{
\code{dskewnorm} gives the density, \code{pskewnorm} gives the distribution function, \code{qskewnorm} gives the quantile function, and \code{rskewnorm} generates random deviates.
}
\description{
Density, distribution function, quantile function, and random generation for
the skew normal distribution.
}
\details{
This implementation of \code{dskewnorm} allows for automatic differentiation with \code{RTMB} while the other functions are imported from the \code{sn} package.
When called during AD taping, \code{pskewnorm} instead integrates the density numerically with the AD-compatible \code{\link[RTMB:ADintegrate]{integrate}} of \code{RTMB}, which makes it AD-compatible as well.
See \code{sn::\link[sn]{dsn}} for more details.

\deqn{f(x;\,\xi,\omega,\alpha) = \frac{2}{\omega}\,\phi\!\left(\frac{x-\xi}{\omega}\right)\Phi\!\left(\alpha\frac{x-\xi}{\omega}\right),}
where \eqn{\phi} and \eqn{\Phi} are the standard normal PDF and CDF.
}
\examples{
# alpha is skew parameter
x <- rskewnorm(1, alpha = 1)
d <- dskewnorm(x, alpha = 1)
p <- pskewnorm(x, alpha = 1)
q <- qskewnorm(p, alpha = 1)
}
\seealso{
\link{skewnorm2}, \link{skewt}, \link{skewt2}
}
