% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnorm2.R
\name{skewnorm2}
\alias{skewnorm2}
\alias{dskewnorm2}
\alias{pskewnorm2}
\alias{qskewnorm2}
\alias{rskewnorm2}
\title{Reparameterised skew normal distribution}
\usage{
dskewnorm2(x, mean = 0, sd = 1, alpha = 0, log = FALSE)

pskewnorm2(q, mean = 0, sd = 1, alpha = 0, lower.tail = TRUE, log.p = FALSE)

qskewnorm2(p, mean = 0, sd = 1, alpha = 0, lower.tail = TRUE, log.p = FALSE)

rskewnorm2(n, mean = 0, sd = 1, alpha = 0)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mean}{mean parameter}

\item{sd}{standard deviation, must be positive.}

\item{alpha}{skewness parameter, +/- \code{Inf} is allowed.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/ densities \eqn{p} are returned as \eqn{\log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \le x]}, otherwise \eqn{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of random values to return}
}
\value{
\code{dskewnorm2} gives the density, \code{pskewnorm2} gives the distribution function, \code{qskewnorm2} gives the quantile function, and \code{rskewnorm2} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for
the skew normal distribution reparameterised in terms of mean, standard deviation and skew magnitude
}
\details{
This implementation of \code{dskewnorm2} allows for automatic differentiation with \code{RTMB} while the other functions are imported from the \code{sn} package.
When called during AD taping, \code{pskewnorm2} instead integrates the density numerically with the AD-compatible \code{\link[RTMB:ADintegrate]{integrate}} of \code{RTMB}, which makes it AD-compatible as well.

Uses the same density as \code{skewnorm} with location \eqn{\xi}, scale \eqn{\omega}, and shape \eqn{\alpha}
reparameterised from the mean \eqn{m}, standard deviation \eqn{s}, and skewness parameter \eqn{\alpha}:
\deqn{\delta = \frac{\alpha}{\sqrt{1+\alpha^2}}, \quad \omega = \frac{s}{\sqrt{1 - 2\delta^2/\pi}}, \quad \xi = m - \omega\delta\sqrt{2/\pi}.}
}
\examples{
# alpha is skew parameter
x <- rskewnorm2(1, alpha = 1)
d <- dskewnorm2(x, alpha = 1)
p <- pskewnorm2(x, alpha = 1)
q <- qskewnorm2(p, alpha = 1)
}
\seealso{
\link{skewnorm}, \link{skewt}, \link{skewt2}
}
