% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2.R
\name{t2}
\alias{t2}
\alias{dt2}
\alias{pt2}
\alias{rt2}
\alias{qt2}
\alias{pt.ad}
\title{Student t distribution with location and scale}
\usage{
dt2(x, mu, sigma, df, log = FALSE)

pt2(q, mu, sigma, df, lower.tail = TRUE, log.p = FALSE)

rt2(n, mu, sigma, df)

qt2(p, mu, sigma, df, lower.tail = TRUE, log.p = FALSE)

pt.ad(q, df)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{sigma}{scale parameter, must be positive.}

\item{df}{degrees of freedom, must be positive.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/ densities \eqn{p} are returned as \eqn{\log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \le x]}, otherwise \eqn{P[X > x]}.}

\item{n}{number of random values to return.}

\item{p}{vector of probabilities}
}
\value{
\code{dt2} gives the density, \code{pt2} gives the distribution function, \code{qt2} gives the quantile function, and \code{rt2} generates random deviates.

\code{pt.ad} is an AD-compatible replacement for \code{stats::pt}. It is used
internally and is reached automatically whenever an argument is an AD variable,
so \code{stats::pt} is left untouched for ordinary use.
}
\description{
Density, distribution function, quantile function, and random generation for
the t distribution with location and scale parameters.
}
\details{
This implementation of \code{dt2} allows for automatic differentiation with \code{RTMB}.

\deqn{f(x;\,\mu,\sigma,\nu) = \frac{1}{\sigma}\,f_t\!\left(\frac{x-\mu}{\sigma};\,\nu\right),}
where \eqn{f_t(\cdot;\nu)} is the Student-\eqn{t} PDF with \eqn{\nu} degrees of freedom.
}
\examples{
x <- rt2(1, 1, 2, 5)
d <- dt2(x, 1, 2, 5)
p <- pt2(x, 1, 2, 5)
q <- qt2(p, 1, 2, 5)
}
