% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztgeom.R
\name{ztgeom}
\alias{ztgeom}
\alias{dztgeom}
\alias{pztgeom}
\alias{rztgeom}
\title{Zero-truncated geometric distribution}
\usage{
dztgeom(x, prob, log = FALSE)

pztgeom(q, prob, lower.tail = TRUE, log.p = FALSE)

rztgeom(n, prob)
}
\arguments{
\item{x, q}{integer vector of counts}

\item{prob}{probability of success in each trial, in (0,1)}

\item{log, log.p}{logical; return log-density if TRUE}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{n}{number of random values to return.}
}
\value{
\code{dztgeom} gives the probability mass function, \code{pztgeom} gives the distribution function, and \code{rztgeom} generates random deviates.
}
\description{
Probability mass function, distribution function, and random generation for
the zero-truncated geometric distribution.
}
\details{
This implementation allows for automatic differentiation with \code{RTMB}.

By definition, this distribution only has support on the positive integers (1, 2, ...).
Any zero-truncated distribution is defined as
\deqn{P(X=x | X>0) = P(X=x) / (1 - P(X=0)),}
where \eqn{P(X=x)} is the probability mass function of the corresponding untruncated distribution.
For the geometric with success probability \eqn{\pi} this gives
\deqn{P(X=x | X>0) = \pi\,(1-\pi)^{x-1}, \quad x = 1, 2, \ldots}
}
\examples{
set.seed(123)
x <- rztgeom(5, prob = 0.3)
d <- dztgeom(x, prob = 0.3)
p <- pztgeom(x, prob = 0.3)
}
\seealso{
\link{zigeom}, \link{hgeom}, \link{ztnbinom}, \link{ztpois}
}
