% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellam.R
\name{skellam}
\alias{skellam}
\alias{dskellam}
\alias{rskellam}
\title{Skellam distribution}
\usage{
dskellam(x, mu1, mu2, log = FALSE)

rskellam(n, mu1, mu2)
}
\arguments{
\item{x}{integer vector of counts}

\item{mu1, mu2}{Poisson means}

\item{log}{logical; return log-density if TRUE}

\item{n}{number of random values to return.}
}
\value{
\code{dskellam} gives the probability mass function and \code{rskellam} generates random deviates.
}
\description{
Probability mass function, distribution function, and random generation for
the Skellam distribution.
}
\details{
The Skellam distribution is the distribution of the difference of two Poisson random variables. Specifically, if \eqn{X_1 \sim \text{Pois}(\mu_1)} and \eqn{X_2 \sim \text{Pois}(\mu_2)}, then \eqn{X_1 - X_2 \sim \text{Skellam}(\mu_1, \mu_2)}.

This implementation of \code{dskellam} allows for automatic differentiation with \code{RTMB}.
}
\examples{
x <- rskellam(1, 2, 3)
d <- dskellam(x, 2, 3)
}
