% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipois.R
\name{zipois}
\alias{zipois}
\alias{dzipois}
\alias{rzipois}
\title{Zero-inflated Poisson distribution}
\usage{
dzipois(x, lambda, zeroprob, log = FALSE)

rzipois(n, lambda, zeroprob)
}
\arguments{
\item{x}{integer vector of counts}

\item{lambda}{vector of (non-negative) means}

\item{zeroprob}{zero-inflation probability between 0 and 1}

\item{log}{logical; return log-density if TRUE}

\item{n}{number of random values to return.}
}
\description{
Zero-inflated Poisson distribution
}
\examples{
set.seed(123)
x = rzipois(1, 0.5, 1)
d = dzipois(x, 0.5, 1)
}
