% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziweibull.R
\name{ziweibull}
\alias{ziweibull}
\alias{dziweibull}
\alias{pziweibull}
\alias{rziweibull}
\title{Zero-inflated Weibull distribution}
\usage{
dziweibull(x, shape, scale, zeroprob = 0, log = FALSE)

pziweibull(q, shape, scale, zeroprob = 0, lower.tail = TRUE, log.p = FALSE)

rziweibull(n, shape, scale, zeroprob = 0)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{shape}{positive shape parameter}

\item{scale}{positive scale parameter}

\item{zeroprob}{zero-inflation probability between 0 and 1.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/ densities \eqn{p} are returned as \eqn{\log(p)}.}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{n}{number of random values to return}
}
\value{
\code{dziweibull} gives the density, \code{pziweibull} gives the distribution function, and \code{rziweibull} generates random deviates.
}
\description{
Density, distribution function, and random generation for
the zero-inflated Weibull distribution.
}
\details{
This implementation allows for automatic differentiation with \code{RTMB}.
}
\examples{
x <- rziweibull(1, 1, 1, 0.5)
d <- dziweibull(x, 1, 1, 0.5)
p <- pziweibull(x, 1, 1, 0.5)
}
